requirements.txt
selenium==3.8.0
beautifulsoup4
requests
imgurpython
pyimgur
line-bot-sdk
oauth2client
flask_sqlalchemy
flask_restful
flask_cors
Flask
pymysql
python-dotenv
pathlib
gunicorn
gspread
pandas
datetime
matplotlib